feature: flagging unrevealed cells
user can right-click on a button and a flag will pop up, allowing you to flag cells you think have a bomb. NOTE that in the bomb map, we don't care about checking that toggleFlag is being called for an unrevealed cell. In the front end, only unrevealed cells have buttons that show up.